requirements.txt
discord.py==2.5.2
openai==1.75.0
anthropic==0.51.0
google-api-python-client==2.167.0
google-auth==2.39.0
